pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-drive"
version = "0.1.0"
authors = [
  { name="VanAI Team", email="hello@vanaipower.com" },
]
description = "The Enterprise File System and Workspace for Autonomous AI Agents."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "azure-identity>=1.15.0",
    "PyJWT>=2.8.0",
    # "azure-storage-file-share>=12.13.0" # Required for AzureProvider
]

[project.urls]
"Homepage" = "https://github.com/lixinso/agent-drive"
"Bug Tracker" = "https://github.com/lixinso/agent-drive/issues"
